backend/requirements.txt
numpy==2.1.3
PyMySQL
requests~=2.32.3
scipy
SQLAlchemy
pandas
matplotlib
python-dotenv
akshare
tushare
yfinance
typing-extensions
langgraph
chromadb
dashscope
langchain-openai
langchain_anthropic
langchain-google-genai
tqdm
langchain-mcp-adapters
pydantic
openai
fastmcp
beautifulsoup4~=4.13.3
mcp
tenacity~=9.0.0
efinance~=0.5.5.2
fastapi
uvicorn
mplfinance
TA-Lib
langchain-core
a2a-sdk[http-server]
httpx
